fix Tournament機能をpongからtournamentへ分離(frontend)
