+++ django/backend/ft_trans/settings_dev.py
@@ -99,6 +99,7 @@ INSTALLED_APPS = [
     "spa",
     "notification",
     "pong",
+    "tournament",
     "friend",
     # "login",
     "accounts",
