+++ django/backend/ft_trans/settings_prod.py
@@ -98,6 +98,7 @@ INSTALLED_APPS = [
     "spa",
     "notification",
     "pong",
+    "tournament",
     "friend",
     # "login",
     "accounts",
