+++ django/backend/ft_trans/urls.py
@@ -18,12 +18,14 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import include, path, re_path
 from django.conf.urls.i18n import i18n_patterns
-import pong.urls
+
 
 import spa.urls
 import accounts.urls
 import friend.urls
 import notification.urls
+import pong.urls
+import tournament.urls
 
 
 # API
@@ -38,6 +40,7 @@ urlpatterns += i18n_patterns(
     path("notification/", include(notification.urls), name="notification"),
     path("accounts/", include(accounts.urls), name="accounts"),
     path("pong/", include(pong.urls), name="pong"),
+    path("tournament/", include(tournament.urls), name="tournament"),
     path("friend/", include(friend.urls), name="friend"),
     path("spa/", include(spa.urls), name="spa2"),
     re_path(r"[\w\-\/]*", include(spa.urls), name="spa"),
