--- django/backend/pong/templates/pong/card.html
@@ -1,58 +1,0 @@
-{% load i18n %}
-<body>
-  <br />
-  <ul>
-    <div class="container mt-4">
-      <div class="d-flex align-items-end">
-        <h3>{{attr.title}}</h3>
-        <span class="h6 ps-2">
-          <a href="{{attr.link}}" class="">{% trans 'すべての項目を表示' %}</a>
-        </span>
-      </div>
-
-      <div class="row">
-        {% for tournament in tournaments %}
-
-        <div class="col-md-4 mb-4">
-          <div class="card">
-            <div class="card-header">
-              <h5 class="card-title text-center">{{tournament.name}}</h5>
-            </div>
-            <div class="card-body text-truncate">
-              <div class="text-center">
-                <strong class="card-label d-flex justify-content-center"
-                  >{% trans '主催者'%}<br /></strong
-                ><span><span class="text-center">{{tournament.organizer.username}}</span></span>
-              </div>
-              <div class="text-center mt-2">
-                <strong class="card-label d-flex justify-content-center"
-                  >{% trans '最大参加人数'%}<br /></strong
-                ><span class="d-flex justify-content-center">{{tournament.current_players}}</span>
-              </div>
-              <div class="text-center mt-2">
-                <strong class="card-label d-flex justify-content-center"
-                  >{% trans '開始日時'%}<br /></strong
-                ><span class="d-flex justify-content-center">{{tournament.start_at}}</span>
-              </div>
-            </div>
-
-            <form
-              id="register-tournament"
-              method="post"
-              action="{% url 'pong:register-tournament'  %}"
-            >
-              {% csrf_token %}
-              <input name="id" type="hidden" value="{{tournament.id}}" />
-              <input name="status" type="hidden" value="ACCEPTED" />
-              <div class="card-footer text-center">
-                <button type="submit" class="btn btn-primary">{{attr.button}}</button>
-              </div>
-            </form>
-          </div>
-        </div>
-
-        {% endfor %}
-      </div>
-    </div>
-  </ul>
-</body>
