+++ django/backend/tournament/apps.py
@@ -1,0 +1,12 @@
+from django.apps import AppConfig
+
+
+class TournamentConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "tournament"
+
+    def ready(self):
+        import tournament.signals
+
+        # Black対策
+        print(f"f{tournament.signals}")
