+++ django/backend/tournament/forms.py
@@ -1,7 +1,13 @@
 from django import forms
 from .models import Tournament
 from django.utils.translation import gettext_lazy as _
-from datetime import datetime, timezone, timedelta
+from datetime import datetime, timezone
+
+
+TIME_HOUR_CHOICES = [((f"{hour:02d}"), f"{hour:02d}") for hour in range(0, 24)]
+TIME_MINUTE_CHOICES = [
+    ((f"{minute:02d}"), f"{minute:02d}") for minute in [0, 15, 30, 45]
+]
 
 
 class TournamentForm(forms.ModelForm):
@@ -11,15 +17,38 @@ class TournamentForm(forms.ModelForm):
                 "id": "start-datetime",
                 "class": "form-control datetimepicker-input text-black",
                 "verbose_name": _("最大参加人数"),
-                "placeholder": _("トーナメント開始日時"),
-                "type": "datetime-local",
+                "placeholder": _("トーナメント開始日"),
+                "type": "date",
                 "value": (
-                    datetime.now(tz=timezone.utc) + timedelta(seconds=14400)
-                ).strftime("%Y-%m-%dT%H:00"),
-                "step": "300",
+                    # datetime.now(tz=timezone.utc) + timedelta(seconds=14400)
+                    datetime.now(tz=timezone.utc)
+                ).strftime("%Y-%m-%d"),
+                # ).strftime("%Y-%m-%dT%H:00"),
             }
         ),
     )
+    # start_time = forms.ChoiceField(
+    #    choices=TIME_CHOICES,
+    # )
+
+    start_hour = forms.ChoiceField(
+        choices=TIME_HOUR_CHOICES,
+        widget=forms.Select(attrs={"class": "dropdown-toggle  btn border text-black"}),
+    )
+    start_minute = forms.ChoiceField(
+        choices=TIME_MINUTE_CHOICES,
+        widget=forms.Select(attrs={"class": "dropdown-toggle btn border text-black"}),
+    )
+    # start_hour = forms.IntegerField(
+    #    min_value=0,
+    #    max_value=23,
+    # )
+    # start_minute = forms.IntegerField(
+    #    min_value=0,
+    #    max_value=45,
+    #    step_size=15,
+    # )
+
     name = forms.CharField(
         max_length=32,
         widget=forms.TextInput(
@@ -54,6 +83,8 @@ class TournamentForm(forms.ModelForm):
             "is_only_friend",
             "current_players",
             "start_at",
+            "start_hour",
+            "start_minute",
         )
 
     def set_organizer(self, name):
