+++ django/backend/tournament/signals.py
