+++ django/backend/tournament/templates/tournament/card.html
@@ -1,0 +1,80 @@
+{% load i18n %}
+<body>
+  <br />
+  <ul>
+    <div class="container mt-4">
+      <div class="d-flex align-items-center align-items-sm-end flex-column flex-sm-row">
+        <h3>{{attr.title}}</h3>
+
+        {% if tournaments %}
+        <span class="h6 ps-2">
+          <a href="{{attr.link}}" class="">{% trans 'すべての項目を表示' %}</a>
+        </span>
+        {% endif %}
+      </div>
+
+      {% if tournaments %}
+      <div class="row d-flex justify-content-center justify-content-sm-between overflow-hidden">
+        {% for tournament in tournaments %}
+
+        <div class="col-md-4 mb-4 w-auto">
+          <div class="card mx-0 mx-sm-0">
+            <div class="card-header">
+              <h5 class="card-title text-center">{{tournament.name}}</h5>
+            </div>
+            <div class="card-body text-truncate">
+              <div class="text-center">
+                <strong class="card-label d-flex justify-content-center"
+                  >{% trans '主催者'%}<br /></strong
+                ><span><span class="text-center">{{tournament.organizer.username}}</span></span>
+              </div>
+              <div class="text-center mt-3">
+                <strong class="card-label d-flex justify-content-center"
+                  >{% trans '最大参加人数'%}<br /></strong
+                ><span class="d-flex justify-content-center">{{tournament.current_players}}</span>
+              </div>
+              <div class="text-center mt-3">
+                <strong class="card-label d-flex justify-content-center"
+                  >{% trans '開始日時'%}<br /></strong
+                ><span class="d-flex justify-content-center">{{tournament.start_at}}</span>
+              </div>
+            </div>
+
+            <div class="d-flex justify-content-center w-100">
+              {% if attr.display_register %}
+              <form
+                id="register-tournament"
+                method="post"
+                action="{% url 'tournament:register'  %}"
+                class="w-100 register"
+              >
+                {% csrf_token %}
+                <input name="id" type="hidden" value="{{tournament.id}}" />
+                <input name="status" type="hidden" value="ACCEPTED" />
+                <div class="card-footer text-center">
+                  <button type="submit" class="btn btn-primary">{{attr.button}}</button>
+                </div>
+              </form>
+              {% endif %}
+
+              <form method="post" action="{% url 'tournament:register'  %}" class="w-100 details">
+                {% csrf_token %}
+                <input name="id" type="hidden" value="{{tournament.id}}" />
+                <input name="status" type="hidden" value="ACCEPTED" />
+                <div class="card-footer text-center">
+                  <button type="submit" class="btn btn-primary">{% trans '詳細 '%}</button>
+                </div>
+              </form>
+            </div>
+          </div>
+        </div>
+
+        {% endfor %}
+      </div>
+
+      {% else %}
+      <div>{% trans '対象のトーナメントは存在しません' %}</div>
+      {% endif %}
+    </div>
+  </ul>
+</body>
