+++ django/backend/tournament/templates/tournament/list.html
