+++ django/frontend/src/tournament/assets/medal-crown-10328-gold.png
