+++ django/frontend/src/tournament/js/rebuild_tournament.js
