+++ django/frontend/src/tournament/js/tournament.js
