+++ django/frontend/src/tournament/js/tournament/Branch.js
