+++ django/frontend/src/tournament/js/tournament/Line.js
