+++ django/frontend/src/tournament/js/tournament/Point.js
