+++ django/frontend/src/tournament/js/tournament/TournamentChart.js
@@ -1,4 +1,4 @@
-import Point from './/Point.js';
+import Point from './Point.js';
 //import Branch from './/Branch.js';
 import Game from './Game.js';
