+++ django/frontend/src/tournament/js/tournament_chart.js
