+++ django/frontend/src/tournament/scss/tournament.scss
@@ -77,6 +77,7 @@
 }
 
 .card {
-  max-width: 20rem; /* 横幅を最大20remに制限 */
-  margin: 0 auto; /* カードを中央揃えにする */
+  max-width: 15rem;
+  min-width: 15rem;
+  //margin: 0 auto;
 }
